Integrate loader to player
